sound ok not close

+++ src/main.py
@@ -1,6 +1,6 @@
 import time, os, re
-from uuid import UUID
-import subprocess
+
+USB_TEST = "082b5a3d-dd9b-478e-8a7d-6c906466e6b3"
 
 devices = os.popen('sudo blkid').readlines()
 
@@ -16,9 +16,11 @@ for u in devices:
 
 print(usbs[0]["PARTUUID"])
 
-commande_sound = 'mpg123 "../data/viead.mp3"'
-commande_sound_closed = "pkill mpg123"
+if usbs[0]["PARTUUID"] == USB_TEST:
 
-process = os.popen(commande_sound)
-time.sleep(10)
-os.popen(commande_sound_closed)
+    commande_sound = 'mpg123 "../data/viead.mp3"'
+    commande_sound_closed = "pkill mpg123"
+
+    process = os.popen(commande_sound)
+    time.sleep(10)
+    os.popen(commande_sound_closed)
